+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # CHANGELOG
+
+## v0.2 - 202605121
+Step 2: try sharding in parallel
+0. copy the config/default_template.yml to config/default.yml, and update values.
+1. copy the template from "jobs" directory and edit the "TODO"s
+2. submit the edited script by calling "sbatch my_edited_script.sh"
+
+The numbers in /jobs & /tests are corresponding. (eg. job 01 calls test 01 etc)
+
 ## v0.1 - 20260519
 Step 1: Prepare the environment. For now just on LUMI using containers and try the imports.